adjusted functions
