+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "eslint": "8.44.0",
     "eslint-config-next": "13.4.8",
+    "nanoid": "^5.0.7",
     "next": "13.4.8",
     "react": "18.2.0",
     "react-dom": "18.2.0",
